requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pandas==2.2.2
python-dotenv==1.0.1
